% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{batt_cols}
\alias{batt_cols}
\title{Battlytics brand colours}
\usage{
batt_cols()
}
\value{
A named character vector of hex colours.
}
\description{
A named vector of Battlytics colours derived from the brand logo.
}
