% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_battlytics.R
\name{batt_labs}
\alias{batt_labs}
\title{Standard Battlytics labels}
\usage{
batt_labs(title, subtitle = NULL, caption = NULL, x = NULL, y = NULL)
}
\arguments{
\item{title}{Plot title.}

\item{subtitle}{Plot subtitle.}

\item{caption}{Plot caption, usually includes source.}

\item{x}{X-axis title.}

\item{y}{Y-axis title.}
}
\value{
A ggplot2 labels object.
}
\description{
Standard Battlytics labels
}
