% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{batt_pal}
\alias{batt_pal}
\title{Battlytics palette}
\usage{
batt_pal(palette = "main", n = 3)
}
\arguments{
\item{palette}{Palette name. Options: "main", "cool".}

\item{n}{Number of colours to return.}
}
\value{
A character vector of hex colours.
}
\description{
Returns a vector of colours for use in charts.
}
