% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_battlytics}
\alias{scale_fill_battlytics}
\title{Battlytics fill scale for ggplot2}
\usage{
scale_fill_battlytics(palette = "main", ...)
}
\arguments{
\item{palette}{Palette name. Options: "main", "cool".}

\item{...}{Passed to ggplot2::discrete_scale}
}
\description{
Battlytics fill scale for ggplot2
}
